udpated BD

+++ Duc.Splitt/Duc.Splitt.Data/DataAccess/Models/BackOfficeUser.cs
@@ -33,14 +33,14 @@ public partial class BackOfficeUser
     [Column(TypeName = "datetime")]
     public DateTime CreatedOn { get; set; }
 
-    public byte CreatedAt { get; set; }
+    public int CreatedAt { get; set; }
 
     public Guid? ModifiedBy { get; set; }
 
     [Column(TypeName = "datetime")]
     public DateTime? ModifiedOn { get; set; }
 
-    public byte? ModifiedAt { get; set; }
+    public int? ModifiedAt { get; set; }
 
     [ForeignKey("CreatedAt")]
     [InverseProperty("BackOfficeUserCreatedAtNavigation")]
